services/vanna/requirements.txt
fastapi
uvicorn[standard]
httpx
psycopg[binary]
python-dotenv
pydantic
